Extract trackjs embed
