+++ src/Twig/MonitoringTwigExtension.php
@@ -2,97 +2,25 @@
 
 namespace Becklyn\Hosting\Twig;
 
-use Becklyn\AssetsBundle\Helper\AssetHelper;
-use Becklyn\Hosting\Config\HostingConfig;
-use Symfony\Component\Asset\Packages;
+use Becklyn\Hosting\TrackJS\TrackJSEmbed;
 
 
 class MonitoringTwigExtension extends \Twig_Extension
 {
     /**
-     * @var HostingConfig
+     * @var TrackJSEmbed
      */
-    private $hostingConfig;
-
-    /**
-     * @var string
-     */
-    private $environment;
-
-    /**
-     * @var string
-     */
-    private $isDebug;
-
-    /**
-     * @var AssetHelper|null
-     */
-    private $assetHelper;
-
-    /**
-     * @var Packages|null
-     */
-    private $packages;
+    private $trackJSEmbed;
 
 
     /**
-     * @param HostingConfig    $hostingConfig
-     * @param AssetHelper|null $assetHelper
-     * @param Packages|null    $packages
-     * @param string           $environment
-     * @param string           $isDebug
+     * @param TrackJSEmbed $trackJSEmbed
      */
-    public function __construct (
-        HostingConfig $hostingConfig,
-        ?AssetHelper $assetHelper,
-        ?Packages $packages,
-        string $environment,
-        string $isDebug
-    )
+    public function __construct (TrackJSEmbed $trackJSEmbed)
     {
-        $this->hostingConfig = $hostingConfig;
-        $this->assetHelper = $assetHelper;
-        $this->packages = $packages;
-        $this->environment = $environment;
-        $this->isDebug = $isDebug;
+        $this->trackJSEmbed = $trackJSEmbed;
     }
 
-
-    /**
-     * @return string
-     */
-    public function embedMonitoring () : string
-    {
-        if (null === $this->assetHelper && null === $this->packages)
-        {
-            throw new AssetIntegrationFailedException("No asset integration extension found. Please either install `becklyn/assets-bundle` or `symfony/asset` to use this bundle.");
-        }
-
-        $trackJsToken = $this->hostingConfig->getTrackJsToken();
-
-        // only embed if token is set, in production and not in debug
-        if (null === $trackJsToken || $this->isDebug || "prod" !== $this->environment)
-        {
-            return "";
-        }
-
-        $assetUrl = null !== $this->assetHelper
-            ? $this->assetHelper->getUrl("@hosting/js/trackjs.js")
-            : $this->packages->getUrl("bundles/becklynhosting/js/trackjs.js");
-
-        return \sprintf(
-            '<script src="%s"></script><script>window.TrackJS && TrackJS.install(%s)</script>',
-            $assetUrl,
-            \json_encode([
-                "token" => $trackJsToken,
-                "application" => $this->hostingConfig->getProjectName(),
-                "version" => $this->hostingConfig->getGitCommit(),
-                "console" => [
-                    "display" => false,
-                ],
-            ])
-        );
-    }
 
     /**
      * @inheritdoc
@@ -100,7 +28,7 @@ class MonitoringTwigExtension extends \Twig_Extension
     public function getFunctions () : iterable
     {
         return [
-            new \Twig_Function("hosting_embed_monitoring", [$this, "embedMonitoring"], ["is_safe" => ["html"]]),
+            new \Twig_Function("hosting_embed_monitoring", [$this->trackJSEmbed, "getEmbedHtml"], ["is_safe" => ["html"]]),
         ];
     }
 }
